requirements.txt
chainlit==2.8.4
fastapi==0.116.1
python-dotenv==1.1.1
uvicorn==0.35.0
langchain==1.0.3
langchain-core==1.0.3
langchain-google-genai==3.0.0